refactor(tmux): split platform-specific font detection via build constraints
The iTerm2 font auto-detection shells to macOS-only tools (`defaults`,
`system_profiler`). Instead of `runtime.GOOS` guards with that code compiled
into every platform, move it behind a build-constrained `detectTerminalFontFile`:
  - tmux_font_darwin.go (//go:build darwin): real iTerm2 detection.
  - tmux_font_other.go  (//go:build !darwin): stub returning a "pass --base" error.
The cross-platform orchestration (locate script, find python, patch) stays in
tmux_install_wizard.go. Verified it cross-compiles for darwin/linux/windows.
